requirements.txt
bottle==0.12.21
requests==2.27.1
tensorflow-cpu
